pyproject.toml
[project]
name = "orchestra"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.5",
    "mysqlclient>=2.2.7",
    "playwright==1.54.0",
    "psycopg2-binary>=2.9.10",
    "pyright>=1.1.403",
    "python-decouple>=3.8",
    "python-dotenv>=1.1.1",
    "ruff>=0.12.10",
    "django-stubs>=5.2.2",
    "filelock>=3.19.1",
    "psutil>=7.0.0",
    "pydantic>=2.6.0,<2.10.0",
    "pydantic-core",
    "bandit>=1.8.6",
    "celery>=5.5.3",
    "celery-stubs>=0.0.2",
    "django-extensions>=4.1",
    "redis>=6.4.0",
]

[tool.ruff]
line-length = 88
target-version = "py312"
preview = true
exclude = [
    ".venv",
    "migrations",
    "__pycache__",
    "htmlcov",
    "media",
    "logs",
    ".direnv",
    ".git",
    ".git-rewrite",
    ".mypy_cache",
    ".ruff_cache",
    "venv",
    "env",
]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "N", "D", "UP", "B", "A", "C4", "SIM", "ISC", "ICN", "TID", "ERA", "RUF", "DJ", "PL", "C90", "ANN", "ASYNC", "PERF", "PIE", "NPY"]
ignore = [
    "E501", # Line too long (handled by formatter)
    "F401", # Unused import (handled by ruff --fix)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint.isort]
known-first-party = ["apps", "core"]
known-local-folder = []
force-single-line = false
combine-as-imports = true

[tool.ruff.lint.pydocstyle]
convention = "google"


[dependency-groups]
dev = [
    "decouple-types>=1.0.2",
    "django-debug-toolbar>=6.0.0",
    "line-profiler==5.0.0",
    "pre-commit>=4.2.0",

    "snakeviz>=2.2.2",
    "scipy-stubs>=1.11.4.1",
]
all = [
    "django>=5.2.5",
    "mysqlclient>=2.2.7",
    "playwright==1.54.0",
    "psycopg2-binary>=2.9.10",
    "pyright>=1.1.403",
    "python-decouple>=3.8",
    "python-dotenv>=1.1.1",
    "ruff>=0.12.10",
    "django-stubs>=5.2.2",
    "filelock>=3.19.1",
    "psutil>=7.0.0",
    "pydantic>=2.6.0,<2.10.0",
    "pydantic-core",
    "bandit>=1.8.6",
    "celery>=5.5.3",
    "celery-stubs>=0.0.2",
    "django-extensions>=4.1",
    "redis>=6.4.0",
    "decouple-types>=1.0.2",
    "django-debug-toolbar>=6.0.0",
    "line-profiler==5.0.0",
    "pre-commit>=4.2.0",

    "snakeviz>=2.2.2",
    "scipy-stubs>=1.11.4.1",
]
